+++ src/main/kotlin/com/springkotlin/springsplit/repositories/UserRepository.kt
@@ -4,7 +4,7 @@ import com.springkotlin.springsplit.entities.User
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface UserRepository:JpaRepository<User,Int> {
+interface UserRepository:JpaRepository<User,Long> {
 
     fun findByEmail(email: String):User?
     fun findByEmailAndPassword(email:String,password:String):User?
